+++ sdk/packages/sdk/src/chains/substrate.ts
@@ -25,7 +25,11 @@ import {
 	replaceWebsocketWithHttp,
 	parseStateMachineId,
 } from "@/utils"
-import { ExpectedError, MissingConsensusUpdateTimeError } from "@/utils/exceptions"
+import {
+	ExpectedError,
+	MissingConsensusUpdateTimeError,
+	MISSING_CONSENSUS_UPDATE_TIME_MESSAGE,
+} from "@/utils/exceptions"
 import { keccakAsU8a } from "@polkadot/util-crypto"
 import { ISMP_PREFIX } from "@/configs/constants"
 
@@ -409,7 +413,7 @@ export class SubstrateChain implements IChain {
 			return BigInt(updateTime)
 		} catch (error) {
 			if (MissingConsensusUpdateTimeError.isError(error)) {
-				throw new MissingConsensusUpdateTimeError(undefined, { cause: error })
+				throw new MissingConsensusUpdateTimeError(MISSING_CONSENSUS_UPDATE_TIME_MESSAGE, { cause: error })
 			}
 			throw error
 		}